chore: make it run under venv on python 3.12

+++ rl_6_nimmt/agents/random.py
@@ -6,7 +6,7 @@ class DrunkHamster(Agent):
     """ Agent that makes totally random decisions """
 
     def forward(self, state, legal_actions, **kwargs):
-        action = np.random.choice(np.array(legal_actions, dtype=np.int), size=1)[0]
+        action = np.random.choice(np.array(legal_actions, dtype=np.int32), size=1)[0]
         return action, {}
 
     def learn(self, *args, **kwargs):
